+++ src/features/source/browse/screens/SourceMangas.tsx
@@ -196,8 +196,10 @@ const useSourceManga = (
         {
             ...pages[pages.length - 1],
             data: {
+                __typename: 'Mutation',
                 ...lastLoadedPage!.data,
                 fetchSourceManga: {
+                    __typename: 'FetchSourceMangaPayload',
                     ...lastLoadedPage!.data!.fetchSourceManga,
                     hasNextPage:
                         pages.length > lastLoadedPageIndex + 1